requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Mail==0.9.1
bcrypt==4.0.1
python-dotenv==1.0.0
gunicorn==20.1.0
email_validator
werkzeug==2.3.7

